+++ client/web/src/repo/RepoContainer.tsx
@@ -27,6 +27,7 @@ import {
     InteractiveSearchProps,
     repoFilterForRepoRevision,
     CopyQueryButtonProps,
+    quoteIfNeeded,
 } from '../search'
 import { RouteDescriptor } from '../util/contributions'
 import { parseBrowserRepoURL } from '../util/url'
@@ -59,6 +60,7 @@ import { browserExtensionInstalled } from '../tracking/analyticsUtils'
 import { InstallBrowserExtensionAlert } from './actions/InstallBrowserExtensionAlert'
 import { IS_CHROME } from '../marketing/util'
 import { useLocalStorage } from '../util/useLocalStorage'
+import { Settings } from '../schema/settings.schema'
 
 /**
  * Props passed to sub-routes of {@link RepoContainer}.
@@ -100,7 +102,7 @@ const RepoPageNotFound: React.FunctionComponent = () => (
 
 interface RepoContainerProps
     extends RouteComponentProps<{ repoRevAndRest: string }>,
-        SettingsCascadeProps,
+        SettingsCascadeProps<Settings>,
         PlatformContextProps,
         TelemetryProps,
         ExtensionsControllerProps,
@@ -251,7 +253,13 @@ export const RepoContainer: React.FunctionComponent<RepoContainerProps> = props 
                         >
                             <MenuDownIcon className="icon-inline" />
                         </button>
-                        <UncontrolledPopover placement="bottom-start" target="repo-popover" trigger="legacy">
+                        <UncontrolledPopover
+                            placement="bottom-start"
+                            target="repo-popover"
+                            trigger="legacy"
+                            hideArrow={true}
+                            popperClassName="border-0"
+                        >
                             <RepositoriesPopover
                                 currentRepo={repoOrError.id}
                                 history={props.history}
@@ -309,7 +317,7 @@ export const RepoContainer: React.FunctionComponent<RepoContainerProps> = props 
         } else {
             let query = searchQueryForRepoRevision(repoName, globbing, revision)
             if (filePath) {
-                query = `${query.trimEnd()} file:${globbing ? filePath : '^' + escapeRegExp(filePath)}`
+                query = `${query.trimEnd()} file:${quoteIfNeeded(globbing ? filePath : '^' + escapeRegExp(filePath))}`
             }
             onNavbarQueryChange({
                 query,
@@ -328,13 +336,20 @@ export const RepoContainer: React.FunctionComponent<RepoContainerProps> = props 
     ])
 
     const isBrowserExtensionInstalled = useObservable(browserExtensionInstalled)
+    const codeHostIntegrationMessaging =
+        (!isErrorLike(props.settingsCascade.final) &&
+            props.settingsCascade.final?.['alerts.codeHostIntegrationMessaging']) ||
+        'browser-extension'
 
     // Browser extension discoverability features (alert, popover for `GoToCodeHostAction)
     const [hasDismissedExtensionAlert, setHasDismissedExtensionAlert] = useLocalStorage(HAS_DISMISSED_ALERT_KEY, false)
     const [hasDismissedPopover, setHasDismissedPopover] = useState(false)
     const [hoverCount, setHoverCount] = useLocalStorage(HOVER_COUNT_KEY, 0)
     const canShowPopover =
-        !hasDismissedPopover && isBrowserExtensionInstalled === false && hoverCount >= HOVER_THRESHOLD
+        !hasDismissedPopover &&
+        isBrowserExtensionInstalled === false &&
+        codeHostIntegrationMessaging === 'browser-extension' &&
+        hoverCount >= HOVER_THRESHOLD
     const showExtensionAlert = useMemo(
         () => isBrowserExtensionInstalled === false && !hasDismissedExtensionAlert && hoverCount >= HOVER_THRESHOLD,
         // Intentionally use useMemo() here without a dependency on hoverCount to only show the alert on the next reload,
@@ -386,7 +401,7 @@ export const RepoContainer: React.FunctionComponent<RepoContainerProps> = props 
         )
     }
 
-    const repoMatchURL = `/${repoOrError.name}`
+    const repoMatchURL = '/' + repoName.split('/').map(encodeURIComponent).join('/')
 
     const context: RepoContainerContext = {
         ...props,
@@ -406,6 +421,7 @@ export const RepoContainer: React.FunctionComponent<RepoContainerProps> = props 
                     isChrome={IS_CHROME}
                     onAlertDismissed={onAlertDismissed}
                     externalURLs={repoOrError.externalURLs}
+                    codeHostIntegrationMessaging={codeHostIntegrationMessaging}
                 />
             )}
             <RepoHeader