+++ cmd/frontend/graphqlbackend/campaigns.go
@@ -176,7 +176,7 @@ type GitBranchChangesetDescriptionResolver interface {
 
 	Commits() []GitCommitDescriptionResolver
 
-	Published() bool
+	Published() campaigns.PublishedValue
 }
 
 type GitCommitDescriptionResolver interface {