+++ cmd/repo-updater/repoupdater/server_test.go
@@ -986,6 +986,7 @@ func testRepoLookup(db *sql.DB) func(t *testing.T, repoStore repos.Store) func(t
 				githubDotComSource *fakeRepoSource
 				gitlabDotComSource *fakeRepoSource
 				assert             repos.ReposAssertion
+				assertDelay        time.Duration
 				err                string
 			}{
 				{
@@ -1208,6 +1209,34 @@ func testRepoLookup(db *sql.DB) func(t *testing.T, repoStore repos.Store) func(t
 					result: &protocol.RepoLookupResult{ErrorNotFound: true},
 					err:    fmt.Sprintf("repository not found (name=%s notfound=%v)", api.RepoName(githubRepository.Name), true),
 				},
+				{
+					name: "Private repos that used to be public should be removed asynchronously",
+					args: protocol.RepoLookupArgs{
+						Repo: api.RepoName(githubRepository.Name),
+					},
+					githubDotComSource: &fakeRepoSource{
+						err: github.ErrNotFound,
+					},
+					stored: []*repos.Repo{githubRepository},
+					result: &protocol.RepoLookupResult{Repo: &protocol.RepoInfo{
+						ExternalRepo: api.ExternalRepoSpec{
+							ID:          "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAA==",
+							ServiceType: extsvc.TypeGitHub,
+							ServiceID:   "https://github.com/",
+						},
+						Name:        "github.com/foo/bar",
+						Description: "The description",
+						VCS:         protocol.VCSInfo{URL: "git@github.com:foo/bar.git"},
+						Links: &protocol.RepoLinks{
+							Root:   "github.com/foo/bar",
+							Tree:   "github.com/foo/bar/tree/{rev}/{path}",
+							Blob:   "github.com/foo/bar/blob/{rev}/{path}",
+							Commit: "github.com/foo/bar/commit/{commit}",
+						},
+					}},
+					assertDelay: time.Second,
+					assert:      repos.Assert.ReposEqual(),
+				},
 			}
 
 			for _, tc := range testCases {
@@ -1277,6 +1306,9 @@ func testRepoLookup(db *sql.DB) func(t *testing.T, repoStore repos.Store) func(t
 					}
 
 					if tc.assert != nil {
+						if tc.assertDelay != 0 {
+							time.Sleep(tc.assertDelay)
+						}
 						rs, err := store.ListRepos(ctx, repos.StoreListReposArgs{})
 						if err != nil {
 							t.Fatal(err)